split routeguide client
